Move CodeVersionsBuilder to MockDescriptors
